\name{iris}
\alias{iris}
\docType{data}
\title{
My data
}
\description{
Really nice flowers!
}
\usage{data("iris")}
\format{
  A data frame with 150 observations on the following 5 variables.
  \describe{
    \item{\code{Sepal.Length}}{a numeric vector}
    \item{\code{Sepal.Width}}{a numeric vector}
    \item{\code{Petal.Length}}{a numeric vector}
    \item{\code{Petal.Width}}{a numeric vector}
    \item{\code{Species}}{a factor with levels \code{setosa} \code{versicolor} \code{virginica}}
  }
}

\examples{
data(iris)
}

\keyword{datasets}
